PEP8 reformat + python formatting
