+++ embedded_server/sensor_camera_module/BMP280.py
@@ -9,5 +9,6 @@ __status__ = "Prototype"
 import os    # standard library
 import sys
 
+
 def measure():
 	return 42
