+++ embedded_server/sensor_camera_module/DHT11.py
@@ -8,10 +8,11 @@ __status__ = "Prototype"
 
 import random
 
+
 class DHT11:
 
-	def measure(self):
-		return 42+random.randint(0,5)
+    def measure(self):
+        return 42 + random.randint(0, 5)
 
-	def clean(self):
-		pass
+    def clean(self):
+        pass
