+++ embedded_server/sensor_camera_module/INA219.py
@@ -6,8 +6,9 @@ __date__ = "11-08-2020"
 __version__ = "1.0.0"
 __status__ = "Prototype"
 
-import os    # standard library
+import os  # standard library
 import sys
 
+
 def measure():
-	return 42
+    return 42
