+++ embedded_server/sensor_camera_module/WYTP10.py
@@ -6,9 +6,9 @@ __date__ = "11-08-2020"
 __version__ = "1.0.0"
 __status__ = "Prototype"
 
-import os    # standard library
+import os  # standard library
 import sys
 
-def run():
-	# DO SOMETHING
-	pass
+
+def measure():
+    return 42
