+++ embedded_server/smart4l_service/backup/globalVar.py
