+++ embedded_server/smart4l_service/backup/smart4l.py
