+++ embedded_server/smart4l_service/backup/smart4l2.py
