+++ embedded_server/smart4l_service/utils.py
@@ -1,5 +1,7 @@
 # -*- coding: utf-8 -*-
+import sys
 from enum import Enum
+
 
 class Status(Enum):
 	START = 'start'
@@ -7,7 +9,9 @@ class Status(Enum):
 	RESTART = 'restart'
 
 
-class Message():
-	def error(self, msg):
+class Message:
+
+	@staticmethod
+	def error(msg):
 		sys.stderr.write("Error : " + msg)
-		#sys.exit(1)
+		# sys.exit(1)
